Refactored JavaScript component MainControls.js to functional TypeScript component MainControls.tsx
